Use unsigned int for values that require bitwise operations.

+++ include/wx/defs.h
@@ -869,11 +869,7 @@ wxALLOW_COMBINING_ENUMS(wxDirection, wxGeometryCentre)
  * Window (Frame/dialog/subwindow/panel item) style flags
  */
 
-/* The cast is needed to avoid g++ -Wnarrowing warnings when initializing
- * values of int type with wxVSCROLL on 32 bit platforms, where its value is
- * greater than INT_MAX.
- */
-#define wxVSCROLL               ((int)0x80000000) /* FIXME */
+#define wxVSCROLL               0x80000000
 #define wxHSCROLL               0x40000000
 #define wxCAPTION               0x20000000
